Added hexToRGB function and cleaned up code
