+++ src/index.js
@@ -1,9 +1,11 @@
 const calculateColor = require('./calculateColor');
 const getFontColor = require('./getFontColor');
-const toHex = require('./getFontColor');
+const toHex = require('./toHex');
+const hexToRGB = require('./hexToRGB');
 
 module.exports = {
     calculateColor: calculateColor,
     getFontColor: getFontColor,
     toHex: toHex,
+    hexToRGB: hexToRGB,
 }
